The first Version of Drawer
Drawer! Is a paint program made by Marvin Johanning. This program was
firtly designed for my 6 year old sister, so that she can learn how to
use the mouse etc. But I will provide the source code of the program for
everybody to use. The program was made with the Ruby Gem Gosu and
TexPlay.